requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.6
dnspython==2.2.1
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
hyperlink==21.0.0
idna==3.3
imgbbpy==0.1.3
incremental==21.3.0
meilisearch==0.18.3
msgpack==1.0.4
msgpack-python==0.5.6
multidict==6.0.2
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==2.10.6
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
wget==3.2
wrapt==1.14.1
yarl==1.7.2
Zenora==0.0.3
zope.interface==5.4.0
